Layers now used to control Collisions between GameObject
Merged Enemy/Player Projectile into one, therefore instead of checking for either Enemy or Player projectile, it just checks for Projectile. Every gameObject should be in their correct layers. Projectiles would not collide with each other anymore although I've never had it happened but to be on the safe side. Player can collide with Enemy ships(to be implemented).